Fixed icon colour on white status bar
